D8TSDT-120: Refactor helper test functions.
