+++ tests/Functional/ContentNewsRenderTest.php
@@ -4,12 +4,7 @@ declare(strict_types = 1);
 
 namespace Drupal\Tests\oe_theme\Functional;
 
-use Behat\Mink\Element\NodeElement;
-use Drupal\Core\Entity\EntityStorageInterface;
-use Drupal\media\MediaInterface;
 use Drupal\oe_content_entity\Entity\CorporateEntityInterface;
-use Drupal\oe_content_entity_contact\Entity\ContactInterface;
-use Drupal\Tests\BrowserTestBase;
 use Drupal\Tests\oe_theme\PatternAssertions\FieldListAssert;
 use Drupal\user\Entity\Role;
 use Drupal\user\RoleInterface;
@@ -17,7 +12,7 @@ use Drupal\user\RoleInterface;
 /**
  * Tests that our News content type renders correctly.
  */
-class ContentNewsRenderTest extends BrowserTestBase {
+class ContentNewsRenderTest extends ContentRenderTestBase {
 
   /**
    * {@inheritdoc}
@@ -187,125 +182,6 @@ class ContentNewsRenderTest extends BrowserTestBase {
     $details_expected_values['items'][2]['body'] = 'African Court of Justice and Human Rights | Centre for the Development of Enterprise';
     $details_html = $details->getHtml();
     $field_list_assert->assertPattern($details_expected_values, $details_html);
-  }
-
-  /**
-   * Gets the entity type's storage.
-   *
-   * @param string $entity_type_id
-   *   The entity type ID to get a storage for.
-   *
-   * @return \Drupal\Core\Entity\EntityStorageInterface
-   *   The entity type's storage.
-   */
-  protected function getStorage(string $entity_type_id): EntityStorageInterface {
-    return \Drupal::entityTypeManager()->getStorage($entity_type_id);
-  }
-
-  /**
-   * Creates media image entity.
-   *
-   * @param string $name
-   *   Name of the image media.
-   *
-   * @return \Drupal\media\MediaInterface
-   *   Media image instance.
-   */
-  protected function createMediaImage(string $name): MediaInterface {
-    // Create file instance.
-    $file = file_save_data(file_get_contents(drupal_get_path('theme', 'oe_theme') . '/tests/fixtures/placeholder.png'), "public://placeholder_$name.png");
-    $file->setPermanent();
-    $file->save();
-
-    $media = $this->getStorage('media')->create([
-      'bundle' => 'image',
-      'name' => "Test image $name",
-      'oe_media_image' => [
-        'target_id' => (int) $file->id(),
-        'alt' => "Alternative text $name",
-      ],
-      'uid' => 0,
-      'status' => 1,
-    ]);
-    $media->save();
-
-    return $media;
-  }
-
-  /**
-   * Creates Contact entity.
-   *
-   * @param string $name
-   *   Entity name. Is used as a parameter for test data.
-   * @param string $bundle
-   *   Entity bundle.
-   * @param int $status
-   *   Entity status.
-   *
-   * @return \Drupal\oe_content_entity_contact\Entity\ContactInterface
-   *   Contact entity.
-   */
-  protected function createContactEntity(string $name, string $bundle, int $status): ContactInterface {
-    // Create image for contact.
-    $media = $this->createMediaImage($name);
-
-    $contact = $this->getStorage('oe_contact')->create([
-      'bundle' => $bundle,
-      'name' => $name,
-      'oe_address' => [
-        'country_code' => 'BE',
-        'locality' => 'Brussels',
-        'address_line1' => "Address $name",
-        'postal_code' => '1001',
-      ],
-      'oe_body' => "Body text $name",
-      'oe_email' => "$name@example.com",
-      'oe_fax' => "Fax number $name",
-      'oe_mobile' => "Mobile number $name",
-      'oe_office' => "Office $name",
-      'oe_organisation' => "Organisation $name",
-      'oe_phone' => "Phone number $name",
-      'oe_press_contact_url' => ['uri' => "http://www.example.com/press_contact_$name"],
-      'oe_social_media' => [
-        [
-          'uri' => "http://www.example.com/social_media_$name",
-          'title' => "Social media $name",
-          'link_type' => 'facebook',
-        ],
-      ],
-      'oe_website' => ['uri' => "http://www.example.com/website_$name"],
-      'oe_image' => [
-        [
-          'target_id' => (int) $media->id(),
-          'caption' => "Caption $name",
-        ],
-      ],
-      'status' => $status,
-    ]);
-
-    return $contact;
-  }
-
-  /**
-   * Asserts featured media field rendering.
-   *
-   * @param \Behat\Mink\Element\NodeElement $rendered_element
-   *   Rendered element.
-   * @param string $name
-   *   Name of the image media.
-   */
-  protected function assertFeaturedMediaField(NodeElement $rendered_element, string $name): void {
-    $figures = $rendered_element->findAll('css', 'figure.ecl-media-container');
-    $this->assertCount(1, $figures);
-
-    // Assert image tag.
-    $image = $figures[0]->find('css', 'img');
-    $this->assertContains("placeholder_$name.png", $image->getAttribute('src'));
-    $this->assertEquals("Alternative text $name", $image->getAttribute('alt'));
-
-    // Assert caption.
-    $caption = $figures[0]->find('css', 'figcaption');
-    $this->assertEquals("Caption $name", $caption->getText());
   }
 
 }
