+++ tests/Functional/ContentRenderTestBase.php
@@ -5,6 +5,7 @@ declare(strict_types = 1);
 namespace Drupal\Tests\oe_theme\Functional;
 
 use Behat\Mink\Element\NodeElement;
+use Drupal\oe_content_entity_contact\Entity\ContactInterface;
 use Drupal\Tests\BrowserTestBase;
 use Drupal\Core\Entity\EntityStorageInterface;
 use Drupal\media\MediaInterface;
@@ -95,6 +96,112 @@ abstract class ContentRenderTestBase extends BrowserTestBase {
     $file_download_link = $file_row->find('css', '.ecl-file__download');
     $this->assertContains("/sample_$name.pdf", $file_download_link->getAttribute('href'));
     $this->assertContains('Download', $file_download_link->getText());
+  }
+
+  /**
+   * Creates media image entity.
+   *
+   * @param string $name
+   *   Name of the image media.
+   *
+   * @return \Drupal\media\MediaInterface
+   *   Media image instance.
+   */
+  protected function createMediaImage(string $name): MediaInterface {
+    // Create file instance.
+    $file = file_save_data(file_get_contents(drupal_get_path('theme', 'oe_theme') . '/tests/fixtures/placeholder.png'), "public://placeholder_$name.png");
+    $file->setPermanent();
+    $file->save();
+
+    $media = $this->getStorage('media')->create([
+      'bundle' => 'image',
+      'name' => "Test image $name",
+      'oe_media_image' => [
+        'target_id' => (int) $file->id(),
+        'alt' => "Alternative text $name",
+      ],
+      'uid' => 0,
+      'status' => 1,
+    ]);
+    $media->save();
+
+    return $media;
+  }
+
+  /**
+   * Asserts featured media field rendering.
+   *
+   * @param \Behat\Mink\Element\NodeElement $rendered_element
+   *   Rendered element.
+   * @param string $name
+   *   Name of the image media.
+   */
+  protected function assertFeaturedMediaField(NodeElement $rendered_element, string $name): void {
+    $figures = $rendered_element->findAll('css', 'figure.ecl-media-container');
+    $this->assertCount(1, $figures);
+
+    // Assert image tag.
+    $image = $figures[0]->find('css', 'img');
+    $this->assertContains("placeholder_$name.png", $image->getAttribute('src'));
+    $this->assertEquals("Alternative text $name", $image->getAttribute('alt'));
+
+    // Assert caption.
+    $caption = $figures[0]->find('css', 'figcaption');
+    $this->assertEquals("Caption $name", $caption->getText());
+  }
+
+  /**
+   * Creates Contact entity.
+   *
+   * @param string $name
+   *   Entity name. Is used as a parameter for test data.
+   * @param string $bundle
+   *   Entity bundle.
+   * @param int $status
+   *   Entity status.
+   *
+   * @return \Drupal\oe_content_entity_contact\Entity\ContactInterface
+   *   Contact entity.
+   */
+  protected function createContactEntity(string $name, string $bundle, int $status): ContactInterface {
+    // Create image for contact.
+    $media = $this->createMediaImage($name);
+
+    $contact = $this->getStorage('oe_contact')->create([
+      'bundle' => $bundle,
+      'name' => $name,
+      'oe_address' => [
+        'country_code' => 'BE',
+        'locality' => 'Brussels',
+        'address_line1' => "Address $name",
+        'postal_code' => '1001',
+      ],
+      'oe_body' => "Body text $name",
+      'oe_email' => "$name@example.com",
+      'oe_fax' => "Fax number $name",
+      'oe_mobile' => "Mobile number $name",
+      'oe_office' => "Office $name",
+      'oe_organisation' => "Organisation $name",
+      'oe_phone' => "Phone number $name",
+      'oe_press_contact_url' => ['uri' => "http://www.example.com/press_contact_$name"],
+      'oe_social_media' => [
+        [
+          'uri' => "http://www.example.com/social_media_$name",
+          'title' => "Social media $name",
+          'link_type' => 'facebook',
+        ],
+      ],
+      'oe_website' => ['uri' => "http://www.example.com/website_$name"],
+      'oe_image' => [
+        [
+          'target_id' => (int) $media->id(),
+          'caption' => "Caption $name",
+        ],
+      ],
+      'status' => $status,
+    ]);
+
+    return $contact;
   }
 
 }
